modified:   src/MMLib.SwaggerForOcelot/Transformation/SwaggerJsonTransformer.cs
Not sure why this part is added but it solves issue #19 by @nkovacic (All paths and definitions are removed from swagger.json)
as far as I can see.
I'll get further in to this when I have more time but for now on, people that experience the same problem could use this patch
for now.